+++ lib/pow/store/backend/mnesia_cache.ex
@@ -102,27 +102,25 @@ defmodule Pow.Store.Backend.MnesiaCache do
   end
 
   @impl Base
-  @spec put(Config.t(), binary(), any()) :: :ok
-  def put(config, key, value) do
-    GenServer.cast(__MODULE__, {:cache, config, key, value, ttl(config)})
+  def put(config, record_or_records) do
+    ttl = ttl!(config)
+
+    GenServer.cast(__MODULE__, {:cache, config, record_or_records, ttl})
   end
 
   @impl Base
-  @spec delete(Config.t(), binary()) :: :ok
   def delete(config, key) do
     GenServer.cast(__MODULE__, {:delete, config, key})
   end
 
   @impl Base
-  @spec get(Config.t(), binary()) :: any() | :not_found
   def get(config, key) do
-    table_get(config, key)
+    table_get(key, config)
   end
 
   @impl Base
-  @spec keys(Config.t()) :: [any()]
-  def keys(config) do
-    table_keys(config)
+  def all(config, match) do
+    table_all(match, config)
   end
 
   # Callbacks
@@ -136,60 +134,72 @@ defmodule Pow.Store.Backend.MnesiaCache do
   end
 
   @impl GenServer
-  @spec handle_cast({:cache, Config.t(), binary(), any(), integer()}, map()) :: {:noreply, map()}
-  def handle_cast({:cache, config, key, value, ttl}, %{invalidators: invalidators} = state) do
-    table_update(config, key, value, ttl)
+  @spec handle_cast({:cache, Config.t(), Base.record() | [Base.record()], integer()}, map()) :: {:noreply, map()}
+  def handle_cast({:cache, config, record_or_records, ttl}, %{invalidators: invalidators} = state) do
+    invalidators =
+      record_or_records
+      |> table_insert(ttl, config)
+      |> Enum.reduce(invalidators, fn {key, _}, invalidators ->
+        append_invalidator(key, invalidators, ttl, config)
+      end)
 
-    invalidators = update_invalidators(config, invalidators, key, ttl)
     refresh_invalidators_in_cluster(config)
 
     {:noreply, %{state | invalidators: invalidators}}
   end
 
-  @spec handle_cast({:delete, Config.t(), binary()}, map()) :: {:noreply, map()}
+  @spec handle_cast({:delete, Config.t(), Base.key() | [Base.key()]}, map()) :: {:noreply, map()}
   def handle_cast({:delete, config, key}, %{invalidators: invalidators} = state) do
-    invalidators = clear_invalidator(invalidators, key)
-    table_delete(config, key)
+    invalidators =
+      key
+      |> table_delete(config)
+      |> clear_invalidator(invalidators)
 
     {:noreply, %{state | invalidators: invalidators}}
   end
 
   @spec handle_cast({:refresh_invalidators, Config.t()}, map()) :: {:noreply, map()}
   def handle_cast({:refresh_invalidators, config}, %{invalidators: invalidators} = state) do
-    clear_invalidators(invalidators)
-
-    {:noreply, %{state | invalidators: init_invalidators(config)}}
+    {:noreply, %{state | invalidators: init_invalidators(config, invalidators)}}
   end
 
   @impl GenServer
-  @spec handle_info({:invalidate, Config.t(), binary()}, map()) :: {:noreply, map()}
+  @spec handle_info({:invalidate, Config.t(), [Base.key()]}, map()) :: {:noreply, map()}
   def handle_info({:invalidate, config, key}, %{invalidators: invalidators} = state) do
-    invalidators = clear_invalidator(invalidators, key)
-    invalidators =
-      config
-      |> fetch(key)
-      |> delete_or_reschedule(config, invalidators)
+    invalidators = delete_or_reschedule(key, invalidators, config)
 
     {:noreply, %{state | invalidators: invalidators}}
   end
 
-  defp delete_or_reschedule(nil, _config, invalidators), do: invalidators
-  defp delete_or_reschedule({key, _value, key_config, expire}, config, invalidators) do
-    case Enum.max([expire - timestamp(), 0]) do
-      0 ->
-        table_delete(config, key)
-
+  defp delete_or_reschedule(key, invalidators, config) do
+    config
+    |> fetch(key)
+    |> case do
+      nil ->
         invalidators
-      ttl ->
-        update_invalidators(key_config, invalidators, key, ttl)
+
+      {_value, expire} ->
+        case Enum.max([expire - timestamp(), 0]) do
+          0 ->
+            key
+            |> table_delete(config)
+            |> clear_invalidator(invalidators)
+
+          ttl ->
+            append_invalidator(key, invalidators, ttl, config)
+        end
     end
   end
 
-  defp update_invalidators(config, invalidators, key, ttl) do
-    invalidators = clear_invalidator(invalidators, key)
-    invalidator  = trigger_ttl(config, key, ttl)
+  defp append_invalidator(key, invalidators, ttl, config) do
+    invalidators = clear_invalidator(key, invalidators)
+    invalidator  = trigger_ttl(key, ttl, config)
 
     Map.put(invalidators, key, invalidator)
+  end
+
+  defp trigger_ttl(key, ttl, config) do
+    Process.send_after(self(), {:invalidate, config, key}, ttl)
   end
 
   defp refresh_invalidators_in_cluster(config) do
@@ -199,27 +209,21 @@ defmodule Pow.Store.Backend.MnesiaCache do
     |> Enum.each(&:rpc.call(&1, GenServer, :cast, [__MODULE__, {:refresh_invalidators, config}]))
   end
 
-  defp clear_invalidators(invalidators) do
-    Enum.reduce(invalidators, invalidators, fn {key, _ref}, invalidators ->
-      clear_invalidator(invalidators, key)
-    end)
-  end
-
-  defp clear_invalidator(invalidators, key) do
+  defp clear_invalidator(key, invalidators) do
     case Map.get(invalidators, key) do
       nil         -> nil
       invalidator -> Process.cancel_timer(invalidator)
     end
 
-    Map.drop(invalidators, [key])
+    Map.delete(invalidators, key)
   end
 
-  defp table_get(config, key) do
+  defp table_get(key, config) do
     config
     |> fetch(key)
     |> case do
-      {_key, value, _config, _expire} -> value
-      nil                             -> :not_found
+      {value, _expire} -> value
+      nil              -> :not_found
     end
   end
 
@@ -229,54 +233,50 @@ defmodule Pow.Store.Backend.MnesiaCache do
     {@mnesia_cache_tab, mnesia_key}
     |> :mnesia.dirty_read()
     |> case do
-      [{@mnesia_cache_tab, ^mnesia_key, {_key, value, config, expire}} | _rest] -> {key, value, config, expire}
-      [] -> nil
+      [{@mnesia_cache_tab, ^mnesia_key, value} | _rest] -> value
+      []                                                -> nil
     end
   end
 
-  defp table_update(config, key, value, ttl) do
-    mnesia_key = mnesia_key(config, key)
-    expire     = timestamp() + ttl
-    value      = {key, value, config, expire}
+  defp table_all(key_match, config) do
+    mnesia_key_match = mnesia_key(config, key_match)
 
-    :mnesia.sync_transaction(fn ->
-      :mnesia.write({@mnesia_cache_tab, mnesia_key, value})
-    end)
+    @mnesia_cache_tab
+    |> :mnesia.dirty_select([{{@mnesia_cache_tab, mnesia_key_match, :_}, [], [:"$_"]}])
+    |> Enum.map(fn {@mnesia_cache_tab, key, {value, _expire}} -> {unwrap(key), value} end)
   end
 
-  defp table_delete(config, key) do
-    mnesia_key = mnesia_key(config, key)
+  defp unwrap([_namespace, key]), do: key
+  defp unwrap([_namespace | key]), do: key
 
-    :mnesia.sync_transaction(fn ->
-      :mnesia.delete({@mnesia_cache_tab, mnesia_key})
-    end)
+  defp table_insert(record_or_records, ttl, config) do
+    expire  = timestamp() + ttl
+    records = List.wrap(record_or_records)
+
+    {:atomic, _result} =
+      :mnesia.sync_transaction(fn ->
+        Enum.map(records, fn {key, value} ->
+          mnesia_key = mnesia_key(config, key)
+          value      = {value, expire}
+
+          :mnesia.write({@mnesia_cache_tab, mnesia_key, value})
+        end)
+      end)
+
+    records
   end
 
-  defp table_keys(config, opts \\ []) do
-    namespace = mnesia_key(config, "")
+  defp table_delete(key, config) do
+    {:atomic, key} =
+      :mnesia.sync_transaction(fn ->
+        mnesia_key = mnesia_key(config, key)
 
-    sync_all_keys()
-    |> Enum.filter(&String.starts_with?(&1, namespace))
-    |> maybe_remove_namespace(namespace, opts)
-  end
+        :mnesia.delete({@mnesia_cache_tab, mnesia_key})
 
-  defp sync_all_keys do
-    {:atomic, keys} = :mnesia.sync_transaction(fn ->
-      :mnesia.all_keys(@mnesia_cache_tab)
-    end)
+        key
+      end)
 
-    keys
-  end
-
-  defp maybe_remove_namespace(keys, namespace, opts) do
-    case Keyword.get(opts, :remove_namespace, true) do
-      true ->
-        start = String.length(namespace)
-        Enum.map(keys, &String.slice(&1, start..-1))
-
-      _ ->
-        keys
-    end
+    key
   end
 
   defp init_mnesia(config) do
@@ -324,7 +324,7 @@ defmodule Pow.Store.Backend.MnesiaCache do
       :ok
     else
       {:error, reason} ->
-        Logger.error("[inspect __MODULE__}] Couldn't join mnesia cluster because: #{inspect reason}")
+        Logger.error("[#{inspect __MODULE__}] Couldn't join mnesia cluster because: #{inspect reason}")
         {:error, reason}
     end
   end
@@ -370,7 +370,7 @@ defmodule Pow.Store.Backend.MnesiaCache do
 
   defp create_table(config) do
     table_opts = Config.get(config, :table_opts, [disc_copies: [node()]])
-    table_def  = Keyword.merge(table_opts, [type: :set])
+    table_def  = Keyword.merge(table_opts, [type: :ordered_set])
 
     case :mnesia.create_table(@mnesia_cache_tab, table_def) do
       {:atomic, :ok}                                   -> :ok
@@ -402,46 +402,69 @@ defmodule Pow.Store.Backend.MnesiaCache do
   end
 
   defp mnesia_key(config, key) do
-    namespace = Config.get(config, :namespace, "cache")
-
-    "#{namespace}:#{key}"
+    [namespace(config) | List.wrap(key)]
   end
 
-  defp init_invalidators(config) do
-    config
-    |> table_keys(remove_namespace: false)
-    |> Enum.map(&init_invalidator(config, &1))
-    |> Enum.reject(&is_nil/1)
-    |> Enum.into(%{})
+  defp namespace(config), do: Config.get(config, :namespace, "cache")
+
+  defp init_invalidators(config, existing_invalidators \\ %{}) do
+    clear_all_invalidators(existing_invalidators)
+
+    {:atomic, invalidators} =
+      :mnesia.sync_transaction(fn ->
+        :mnesia.foldl(fn
+          {@mnesia_cache_tab, key, {_value, expire}}, invalidators when is_list(key) ->
+            ttl = Enum.max([expire - timestamp(), 0])
+
+            key
+            |> unwrap()
+            |> append_invalidator(invalidators, ttl, config)
+
+          # TODO: Remove by 1.1.0
+          {@mnesia_cache_tab, key, {_key, _value, _config, expire}}, invalidators when is_binary(key) and is_number(expire) ->
+            Logger.warn("[#{inspect __MODULE__}] Deleting old record #{inspect key}")
+
+            :mnesia.delete({@mnesia_cache_tab, key})
+
+            invalidators
+
+          {@mnesia_cache_tab, key, _value}, invalidators ->
+            Logger.warn("[#{inspect __MODULE__}] Found unexpected record #{inspect key}, please delete it")
+
+            invalidators
+        end,
+        %{},
+        @mnesia_cache_tab)
+      end)
+
+    invalidators
   end
 
-  defp init_invalidator(_config, key) do
-    {@mnesia_cache_tab, key}
-    |> :mnesia.dirty_read()
-    |> case do
-      [{@mnesia_cache_tab, ^key, {_key_id, _value, _config, nil}} | _rest] ->
-        nil
-
-      [{@mnesia_cache_tab, ^key, {key_id, _value, config, expire}} | _rest] ->
-        ttl = Enum.max([expire - timestamp(), 0])
-
-        {key, trigger_ttl(config, key_id, ttl)}
-
-      [] -> nil
-    end
-  end
-
-  defp trigger_ttl(config, key, ttl) do
-    Process.send_after(self(), {:invalidate, config, key}, ttl)
+  defp clear_all_invalidators(invalidators) do
+    invalidators
+    |> Map.keys()
+    |> Enum.reduce(invalidators, fn key, invalidators ->
+      clear_invalidator(key, invalidators)
+    end)
   end
 
   defp timestamp, do: :os.system_time(:millisecond)
 
-  defp ttl(config) do
+  defp ttl!(config) do
     Config.get(config, :ttl) || raise_ttl_error()
   end
 
   @spec raise_ttl_error :: no_return
   defp raise_ttl_error,
     do: Config.raise_error("`:ttl` configuration option is required for #{inspect(__MODULE__)}")
+
+  # TODO: Remove by 1.1.0
+  @deprecated "Use `put/2` instead"
+  @doc false
+  def put(config, key, value), do: put(config, {key, value})
+
+  # TODO: Remove by 1.1.0
+  @deprecated "Use `all/2` instead"
+  @doc false
+  def keys(config), do: all(config, :_)
 end